+++ connect/api/v1/project/serializers.py
@@ -318,6 +318,34 @@ class CreateChannelSerializer(serializers.Serializer):
     channeltype_code = serializers.CharField(required=True)
 
 
+class DestroyClassifierSerializer(serializers.Serializer):
+
+    uuid = serializers.CharField(required=True)
+    user_email = serializers.CharField(required=True)
+
+
+class RetrieveClassifierSerializer(serializers.Serializer):
+    uuid = serializers.CharField(required=True)
+
+
+class CreateClassifierSerializer(serializers.Serializer):
+    name = serializers.CharField(required=True)
+    access_token = serializers.CharField(required=True)
+    user = serializers.CharField(write_only=True)
+    project_uuid = serializers.UUIDField(write_only=True)
+
+
+class ClassifierSerializer(serializers.Serializer):
+    project_uuid = serializers.UUIDField()
+
+    def validate_project_uuid(self, value):
+        try:
+            Project.objects.get(uuid=value)
+        except Project.DoesNotExist:
+            raise serializers.ValidationError("This project does not exist")
+        return value
+
+
 class TemplateProjectSerializer(serializers.ModelSerializer):
     class Meta:
         model = ProjectAuthorization