+++ connect/celery.py
@@ -13,6 +13,19 @@ app.config_from_object("django.conf:settings", namespace="CELERY")
 
 app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
 
+task_create_missing_queues = True
+
+# Billing Tasks
+
+app.conf.task_routes = {
+    # 'sync_contacts': {'queue': 'billing'},
+    # 'count_contacts': {'queue': 'billing'},
+    # 'retry_billing_tasks': {'queue': 'billing'},
+    # 'create_contacts': {'queue': 'billing'}
+    # 'get_messages': {'queue': 'billing'},
+}
+
+
 app.conf.beat_schedule = {
     "check-status-services": {
         "task": "connect.common.tasks.status_service",
@@ -52,27 +65,27 @@ app.conf.beat_schedule = {
     },
     "sync-active-contacts": {
         "task": "connect.common.tasks.sync_active_contacts",
-        "schedule": schedules.crontab(hour="*/6")
+        "schedule": schedules.crontab(hour="*/6", minute=0)
     },
     "capture_invoice": {
         "task": "connect.common.tasks.capture_invoice",
         "schedule": schedules.crontab(hour="8,10,13,15,17", minute=0),
     },
-    "sync_contacts": {
-        "task": "sync_contacts",
-        "schedule": schedules.crontab(hour="*/5")
-    },
-    "count_contacts": {
-        "task": "count_contacts",
-        "schedule": schedules.crontab(hour="*/6")
-    },
-    "retry_billing_tasks": {
-        "task": "retry_billing_tasks",
-        "schedule": schedules.crontab(hour="1")
-    },
+    # "sync_contacts": {
+    #     "task": "sync_contacts",
+    #     "schedule": schedules.crontab(hour=settings.SYNC_CONTACTS_SCHEDULE, minute=0)
+    # },
+    # "count_contacts": {
+    #     "task": "count_contacts",
+    #     "schedule": schedules.crontab(hour="*/6", minute=0)
+    # },
+    # "retry_billing_tasks": {
+    #     "task": "retry_billing_tasks",
+    #     "schedule": schedules.crontab(hour="1")
+    # },
     "problem_capture_invoice": {
         "task": "problem_capture_invoice",
-        "schedule": schedules.crontab(hour="9,11,14,16,18")
+        "schedule": schedules.crontab(hour="9,11,14,16,18", minute=0)
     }
 }