+++ connect/common/tasks.py
@@ -55,8 +55,8 @@ def status_service() -> None:
 
 @app.task(name="delete_organization")
 def delete_organization(inteligence_organization: int, user_email):
-    grpc_instance = utils.get_grpc_types().get("inteligence")
-    grpc_instance.delete_organization(
+    ai_client = IntelligenceRESTClient()
+    ai_client.delete_organization(
         organization_id=inteligence_organization,
         user_email=user_email,
     )
@@ -64,26 +64,24 @@ def delete_organization(inteligence_organization: int, user_email):
 
 
 @app.task(name="update_organization")
-def update_organization(inteligence_organization: int, organization_name: str):
-    grpc_instance = utils.get_grpc_types().get("inteligence")
-    grpc_instance.update_organization(
+def update_organization(inteligence_organization: int, organization_name: str, user_email: str):
+    ai_client = IntelligenceRESTClient()
+    ai_client.update_organization(
         organization_id=inteligence_organization,
         organization_name=organization_name,
+        user_email=user_email
     )
     return True
 
 
 @app.task(
-    name="update_user_permission_organization",
-    autoretry_for=(_InactiveRpcError, Exception),
-    retry_kwargs={"max_retries": 5},
-    retry_backoff=True,
+    name="update_user_permission_organization"
 )
 def update_user_permission_organization(
     inteligence_organization: int, user_email: str, permission: int
 ):
-    grpc_instance = utils.get_grpc_types().get("inteligence")
-    grpc_instance.update_user_permission_organization(
+    ai_client = IntelligenceRESTClient()
+    ai_client.update_user_permission_organization(
         organization_id=inteligence_organization,
         user_email=user_email,
         permission=permission,
@@ -146,9 +144,9 @@ def update_user_permission_project(
 @app.task(name="migrate_organization")
 def migrate_organization(user_email: str):
     user = User.objects.get(email=user_email)
-    grpc_instance = utils.get_grpc_types().get("inteligence")
+    ai_client = IntelligenceRESTClient()
 
-    organizations = grpc_instance.list_organizations(user_email=user_email)
+    organizations = ai_client.list_organizations(user_email=user_email)
 
     for organization in organizations:
         org, created = Organization.objects.get_or_create(
@@ -156,8 +154,9 @@ def migrate_organization(user_email: str):
             defaults={"name": organization.get("name"), "description": ""},
         )
 
-        role = grpc_instance.get_user_organization_permission_role(
-            user_email=user_email, organization_id=organization.get("id")
+        role = ai_client.get_user_organization_permission_role(
+            user_email=user_email,
+            organization_id=organization.get("id")
         )
 
         org.authorizations.create(user=user, role=role)
@@ -165,9 +164,8 @@ def migrate_organization(user_email: str):
 
 @app.task(name="create_organization")
 def create_organization(organization_name: str, user_email: str):
-    grpc_instance = utils.get_grpc_types().get("inteligence")
-
-    organization = grpc_instance.create_organization(
+    ai_client = IntelligenceRESTClient()
+    organization = ai_client.create_organization(
         organization_name=organization_name,
         user_email=user_email,
     )
@@ -237,7 +235,7 @@ def update_user_language(user_email: str, language: str):
         user_email=user_email,
         language=language,
     )
-    utils.get_grpc_types().get("inteligence").update_language(
+    IntelligenceRESTClient().update_language(
         user_email=user_email,
         language=language,
     )
@@ -255,10 +253,9 @@ def search_project(organization_id: int, project_uuid: str, text: str):
         )
     )
     inteligence_result = (
-        utils.get_grpc_types()
-        .get("inteligence")
-        .get_organization_inteligences(
-            inteligence_name=text,
+        IntelligenceRESTClient().get_organization_intelligences(
+            intelligence_name=text,
+            organization_id=organization_id
         )
     )
     return {
@@ -358,7 +355,7 @@ def sync_project_statistics():
 @app.task()
 def sync_repositories_statistics():
     flow_instance = utils.get_grpc_types().get("flow")
-    inteligence_instance = utils.get_grpc_types().get("inteligence")
+    ai_client = IntelligenceRESTClient()
 
     for project in Project.objects.all():
         classifiers_project = flow_instance.get_classifiers(
@@ -368,7 +365,7 @@ def sync_repositories_statistics():
         )
         try:
             intelligence_count = int(
-                inteligence_instance.get_count_inteligences_project(
+                ai_client.get_count_intelligences_project(
                     classifiers=classifiers_project,
                 ).get("repositories_count")
             )
@@ -616,6 +613,71 @@ def create_wac_channel(user, flow_organization, config, phone_number_id):
         )
     except grpc.RpcError as error:
         raise error
+
+
+@app.task(name="retrieve_classifier")
+def retrieve_classifier(classifier_uuid: str):
+    grpc_instance = utils.get_grpc_types().get("flow")
+    response = grpc_instance.get_classifier(
+        classifier_uuid=str(classifier_uuid),
+    )
+    return dict(
+        authorization_uuid=response.get("access_token"),
+        classifier_type=response.get("classifier_type"),
+        name=response.get("name"),
+        is_active=response.get("is_active"),
+        uuid=response.get("uuid"),
+    )
+
+
+@app.task(name="destroy_classifier")
+def destroy_classifier(classifier_uuid: str, user_email: str):
+    grpc_instance = utils.get_grpc_types().get("flow")
+    grpc_instance.delete_classifier(
+        classifier_uuid=str(classifier_uuid),
+        user_email=str(user_email),
+    )
+    return True
+
+
+@app.task(name="create_classifier")
+def create_classifier(project_uuid: str, user_email: str, classifier_name: str, access_token):
+    grpc_instance = utils.get_grpc_types().get("flow")
+    response = grpc_instance.create_classifier(
+        project_uuid=project_uuid,
+        user_email=user_email,
+        classifier_type="bothub",
+        classifier_name=classifier_name,
+        access_token=access_token,
+    )
+    return dict(
+        authorization_uuid=response.access_token,
+        classifier_type=response.classifier_type,
+        name=response.name,
+        is_active=response.is_active,
+        uuid=response.uuid,
+    )
+
+
+@app.task(name='list_classifier')
+def list_classifier(project_uuid: str):
+    grpc_instance = utils.get_grpc_types().get("flow")
+    response = grpc_instance.get_classifiers(
+        project_uuid=str(project_uuid),
+        classifier_type="bothub",
+        is_active=True,
+    )
+
+    classifiers = {"data": []}
+    for i in response:
+        classifiers["data"].append({
+            "authorization_uuid": i.get("authorization_uuid"),
+            "classifier_type": i.get("classifier_type"),
+            "name": i.get("name"),
+            "is_active": i.get("is_active"),
+            "uuid": i.get("uuid"),
+        })
+    return classifiers
 
 
 @app.task(name="whatsapp_demo_integration")