+++ connect/utils.py
@@ -39,18 +39,15 @@ def es_convert_datetime(before: str, after: str):
 
 
 def count_contacts(project: Project, before: str, after: str):
-    contacts_day_count = ContactCount.objects.filter(
-        channel__project=project,
-        created_at__lte=before,
-        created_at__gte=after
-    )
+    contacts_day_count = ContactCount.objects.filter(project=project).filter(created_at__range=(after, before))
     return sum([day_count.count for day_count in contacts_day_count])
 
 
 def check_module_permission(claims, user):
     from django.contrib.auth.models import Permission
     from django.contrib.contenttypes.models import ContentType
-    from connect.authentication.models import User
+    from django.contrib.auth import get_user_model
+    User = get_user_model()
 
     if claims.get("can_communicate_internally", False):
         content_type = ContentType.objects.get_for_model(User)