+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 1.2.1
+## Change
+  - Change `SyncManagerTask filter` in `count_contacts`  to only get objects with `status=True` 
+  - remove `task.wait()` from `retry_billing_tasks`
+  - Replaces Message `get_or_create` method with `try/except`
+
+
+# 1.2.0
+## *Add*
+  - Module permission class
+    - On receive a request check if is a module
+  - Internal communication class using rest
+    - class with the generic request from connect to any weni module
+  - New integrations rest endpoints
+    - create endpoints on organization and project view to communicate with integrations
+    - create endpoints and tasks to get integrations information with rest
+
+# 1.1.2
+## CHANGE:
+  - fix `get_contact_detailed` method to get all documents instead of paginated results
+
 # 1.1.1
 ## ADD:
   - exclude support users from authorization list